[ ADD ] Create RecoveryRouter
