+++ src/pages/ChangePasswordPage.tsx
@@ -1,0 +1,5 @@
+export const ChangePasswordPage = () => {
+  return (
+    <div>ChangePasswordPage</div>
+  )
+}
