+++ src/pages/SendEmailRecoveryPage.tsx
@@ -1,0 +1,5 @@
+export const SendEmailRecoveryPage = () => {
+  return (
+    <div>SendEmailRecoveryPage</div>
+  )
+}
