+++ src/pages/index.ts
@@ -1,3 +1,6 @@
-export * from './LandingPage'
-export * from './LoginPage'
-export * from './MainAppPage'
+export * from "./LandingPage";
+export * from "./LoginPage";
+export * from "./MainAppPage";
+export * from "./RegisterPage";
+export * from "./SendEmailRecoveryPage";
+export * from "./ChangePasswordPage";
